+++ src/http/header/if_none_match.rs
@@ -1,6 +1,6 @@
 use super::{EntityTag, IF_NONE_MATCH};
 
-crate::header! {
+crate::__define_common_header! {
     /// `If-None-Match` header, defined in
     /// [RFC7232](https://tools.ietf.org/html/rfc7232#section-3.2)
     ///
@@ -55,11 +55,11 @@ crate::header! {
     (IfNoneMatch, IF_NONE_MATCH) => {Any / (EntityTag)+}
 
     test_if_none_match {
-        test_header!(test1, vec![b"\"xyzzy\""]);
-        test_header!(test2, vec![b"W/\"xyzzy\""]);
-        test_header!(test3, vec![b"\"xyzzy\", \"r2d2xxxx\", \"c3piozzzz\""]);
-        test_header!(test4, vec![b"W/\"xyzzy\", W/\"r2d2xxxx\", W/\"c3piozzzz\""]);
-        test_header!(test5, vec![b"*"]);
+        crate::__common_header_test!(test1, vec![b"\"xyzzy\""]);
+        crate::__common_header_test!(test2, vec![b"W/\"xyzzy\""]);
+        crate::__common_header_test!(test3, vec![b"\"xyzzy\", \"r2d2xxxx\", \"c3piozzzz\""]);
+        crate::__common_header_test!(test4, vec![b"W/\"xyzzy\", W/\"r2d2xxxx\", W/\"c3piozzzz\""]);
+        crate::__common_header_test!(test5, vec![b"*"]);
     }
 }